% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_to_full}
\alias{add_to_full}
\title{Extend a full-row-rank matrix to a full-rank square matrix.}
\usage{
add_to_full(N)
}
\arguments{
\item{N}{A full-row-rank matrix.}
}
\value{
A full-rank square matrix whose first \code{r} rows come from \code{N}.
}
\description{
Given any full-row-rank matrix \code{N} with \code{n=ncol(N)} and \code{r=rank(N)}. 
This function adds \code{n-r} independent rows to the matrix \code{N}, 
such that the resulting new matrix is a full-rank square matrix.
}
\examples{
# generate a full-row-rank matrix
N = matrix(rnorm(10), nrow = 2)

# add 3 independent rows to the matrix to form a full-rank square matrix
add_to_full(N)
}
