% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ind_rows}
\alias{ind_rows}
\title{Select independent rows of a matrix}
\usage{
ind_rows(M)
}
\arguments{
\item{M}{a matrix.}
}
\value{
The function \code{ind_rows} returns a list containing the following components:
  \item{N}{a matrix with \code{r=rank(M)} rows. Its rows come from the selected independent rows in \code{M}.}
  \item{idx}{a logical vector. An element is \code{TRUE} if the corresponding row in \code{M} is selected, \code{FALSE} otherwise.}
}
\description{
Given any matrix \code{M}, this function selects the first \code{r=rank(M)} 
independent rows in \code{M} which form a row basis of \code{M}.
}
\examples{
# generate a matrix M of rank 2 with dependent rows
M = matrix(rnorm(16), nrow = 4)
M[2,] = rnorm(1) * M[1,] + rnorm(1) * M[3,]
M[4,] = rnorm(1) * M[2,] + rnorm(1) * M[3,]

# select the first 2 independent rows
ind_rows(M)
}
