% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polysf.R
\docType{package}
\name{polysf-package}
\alias{polysf-package}
\title{polysf: Strictly feasible representation for a polytope}
\description{
Provide a strictly feasible representation for a polytope with implicit
equality constraints, and a sampler on the degenerated polytope. The same functionality
for Lasso solutions in a non-uniqueness regime, as a special case of the degenerated polytope,
is provided as well.
}
\references{
Telgen, J. (1982). “Minimal representation of convex polyhedral sets”. Journal of
Optimization Theory and Applications 38.1, pp. 1–24.

Drusvyatskiy, Dmitriy, Henry Wolkowicz, et al. (2017). “The many faces of degeneracy
in conic optimization”. Foundations and Trends® in Optimization 3.2, pp. 77–170.

Tibshirani, Ryan J (2013). “The lasso problem and uniqueness”. Electronic Journal
of Statistics 7, pp. 1456–1490.
}
\author{
Xiaozhu Zhang
}
