% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_rep.R
\name{print.sf_rep}
\alias{print.sf_rep}
\title{Print the results of sf_rep}
\usage{
\method{print}{sf_rep}(x, ...)
}
\arguments{
\item{x}{an S3 object of class \code{sf_rep}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{print}, summarizing the implicit equality identification,
and the intrinsic dimension of \code{x}.
}
\description{
The \code{print} method for class \code{sf_rep}.
}
