% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_rep_lasso.R
\name{print.sf_rep_lasso}
\alias{print.sf_rep_lasso}
\title{Print the results of sf_rep_lasso}
\usage{
\method{print}{sf_rep_lasso}(x, ...)
}
\arguments{
\item{x}{an S3 object of class \code{sf_rep_lasso}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{print}, summarizing the equi-correlation set,
the sign of the Theta set, and the intrinsic dimension of \code{x}.
}
\description{
The \code{print} method for class \code{sf_rep_lasso}.
}
