% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_polytope.R
\name{sample_polytope}
\alias{sample_polytope}
\title{Sample uniformly or normally distributed points from a polytope}
\usage{
sample_polytope(obj, npoints, random_walk = NULL, distribution = NULL)
}
\arguments{
\item{obj}{an S3 object of class \code{sf_rep}.}

\item{npoints}{the number of points that the function is going to sample.}

\item{random_walk}{an optional list that declares the random walk and some
related parameters. See the argument \code{random_walk} in \code{volesti::\link[volesti]{sample_points}} 
for details.}

\item{distribution}{an optional list that declares the target density and some 
related parameters. See the argument \code{distribution} in \code{volesti::\link[volesti]{sample_points}}
for details.}
}
\value{
A matrix with \code{npoints} rows and \code{ncol(A)} columns. Each row 
is a sample point from the polytope \eqn{\mathcal{S}}.
}
\description{
This function samples uniformly or normally distributed points
from a polytope \eqn{\mathcal{S} = \{x\in\mathbb{R}^n:Ax\leq b\}} whose 
strictly feasible representation is given by a object of class \code{sf_rep}.
}
\examples{
# generate an S3 object of class sf_poly
A = matrix(c(1, 1, -1, -1, -1, 0, 0, -1), byrow = TRUE, nrow = 4)
b = c(1, -1, 0, 0)
fit = sf_rep(A, b)

# sample 100 uniformly distributed points from Ax<=b
sample_polytope(fit, npoints = 100)

}
